LIB: changed to uint32_t for the printing of the offset

+++ src/CScsi_Non_Medium_Error_Count_Log.cpp
@@ -184,7 +184,7 @@ eReturnValues CScsiNonMediumErrorCountLog::get_Non_Medium_Error_Count_Data(JSONN
 	{
 		JSONNODE *pageInfo = json_new(JSON_NODE);
 		json_set_name(pageInfo, "Non-Medium Error Count Log - 6h");
-		for (size_t offset = 0; offset < m_PageLength; )
+		for (uint32_t offset = 0; offset < m_PageLength; )
 		{
 			if (offset < m_bufferLength && offset < UINT16_MAX)
 			{
